prijs wordt aangepast als helmet true is. daarnaast spdPedals toegevoegd aan de booking. ook is het nu mogelijk om 1 booking op te vragen dmv id booking/{id}

+++ src/main/java/nl/mtbrental/eindproject/model/Booking.java
@@ -17,7 +17,7 @@ public class Booking implements Serializable {
 
     private LocalDateTime date;
 
-    private float amount;
+    private int amount;
 
     @ManyToOne
     private Bike bike;
@@ -79,11 +79,11 @@ public class Booking implements Serializable {
         this.date = date;
     }
 
-    public float getAmount() {
+    public int getAmount() {
         return amount;
     }
 
-    public void setAmount(float amount) {
+    public void setAmount(int amount) {
         this.amount = amount;
     }
